[feature] effective date on working places and policies

+++ app/controllers/api/v1/presence_days_controller.rb
@@ -3,7 +3,6 @@ module API
     class PresenceDaysController < ApplicationController
       authorize_resource except: :create
       include PresenceDayRules
-      include EmployeeBalanceUpdatePresencePerspective
 
       def show
         render_resource(resource)
@@ -29,18 +28,22 @@ module API
 
           transactions do
             resource.update!(attributes)
-            find_and_update_balances(resource.presence_policy) if previous_order != resource.order
+            update_affected_balances(resource.presence_policy) if previous_order != resource.order
           end
           render_no_content
         end
       end
 
       def destroy
-        transactions do
-          find_and_update_balances(resource.presence_policy)
-          resource.destroy!
+        if resource.time_entries.empty?
+          transactions do
+            update_affected_balances(resource.presence_policy)
+            resource.destroy!
+          end
+          render_no_content
+        else
+          locked_error
         end
-        render_no_content
       end
 
       private
